Tracking UPS - Track Sheets
Had to manage numerous shipments and need to keep track of them using a Google Sheet. To automate tracking, I developed a program that uses the Google Sheets API to update the sheet and Selenium to scrape UPS's website for tracking information. By entering "CHECK" in the J column, the program tracks the selected packages and highlights them based on their status.